Fix instrumentation test is terminated unexpectedly
There is a race condition between a timing of when a test process on device
finishes and when AndroidTestListener is notified the test execution complete
because its rely on the timing of Logcat output.

This issue started after I92389c9fb7beb0e87d1fc2b253e5add5a36c8922. Before
that CL, AndroidProcessHandler's terminate status was not tied with a process
on the device and that CL fixed it. With that CL, AndroidProcessHandler's
state becomes terminated upon the termination of the target process. And it
creates this race condition.

This CL fixes the race condition by waiting for test execution Future in
AndroidTestApplicationLaunchTask explicitly. Also this CL adds missing
unit tests for ProcessHandlerLaunchStatus.

Fixes: 140765245
Test: ProcessHandlerLaunchStatusTest
Change-Id: I649e17723cc259df6146a484d41694d0d537a146
(cherry picked from commit 67b0465a6aeaa8888f217bfd4b0b5732bd26b4fc)