[PSD] rename the sample project
Our PSD_SAMPLE project is a grab-bag of features and build script
constructs that are the product of several fevered imaginations.
This is just what we need to torture the new Kotlin build support.

In order to help keep a Kotlin-based and a Groovy-based project in
sync, rename this project to have an explicit GROOVY in its
identifier (and a Groovy/ in its directory path).

Test: existing (no functional change)
Bug: none
Change-Id: I85684d5b0ed4499261fdf0ed3f4722852dafd950