On cold swap, always terminate app first
This CL fixes an issue on cold swap where the user is asked
if they want to disconnect the existing session. This should
always be the case, especially since the button shows the
rerun icon, not the run icon.

It also updates all the tests, as well as remove now-dead
code due to updated assumptions.

Fixes: 142080375
Test: rerun does not prompt the user
Change-Id: I6cf08e86fba3691d0e8cbdf2e8546f31329c1eeb