Include prebuilts m2 folder to dev offline repo.
Artifacts in development like kotlin-compose are placed
under "prebuilts/tools/common/m2/repository", so add this
path to development maven offline repo.

Added flag to RepositoryUrlManager to allow skipping
embedded repo, so that tests in GradleModuleSystemTest
only use the inject repository in testData, and the availabilities
of artifacts are well expected.

Bug: NA
Test: testFindAndroidStudioLocalMavenRepoPaths
Change-Id: Iaa469a105bd65e13674bbc53735274662215afd7