Add color Gutter icons to color declarations in compose
With the flag turned on, you can now see a color gutter icon with
you declare a new Color using the androidx.ui.graphics.Color class.
When you click the color, the same color picker appears as in XML,
you can select a color only if you are using the single int or long
Color constructor.

Test: ComposeColorAnnotatorTest.kt
Bug: 141394302
Change-Id: Id5d77eb6406d73cdce404cc3f0b388bac08305f8