[Dsl,.kts] support expressions with type
It's (relatively) natural to want to set values from extra properties,
but extra properties generically have type Any?, and strongly-typed
setters will therefore need a cast to compile.

Parse an expression with type permissively, using the left-hand side
of the expression.  This is safe for `as`; for `as?`, if the
distinction between the type and null ever becomes relevant to us, we
should inspect the value and the operationReference field of the
expression before deciding what to do.

Test: upcoming AndroidGradleTests
Bug: 130224064
Change-Id: I581dcbf321e830ed8d3b95ce0508056aa98105ee