Use AndroidManifestUtils#getPackageName for Binding classes
Currently, BindingLayout obtains the package name for which we need
to generate binding classes by querying the snapshot from
MergedManifestManager directly. This means that when there are
unrelated manifest merging errors, we have no manifest to read from
and binding class generation fails.

The long-term solution for such issues will be to query the
AndroidManifestIndex and merge the package names of the manifests
that contribute to the module's merged manifest, allowing us to
sidestep any unrelated merging issues.

In the short term, a quick fix for Gradle projects is to use
AndroidManifestUtils#getPackageName instead of using the
MergedManifestManager directly. This API is build system-independent,
and so some implementations will still break due to unrelated merging
issues. For Gradle however, this method simply parses the module's
primary manifest, allowing us to skip merging altogether.

Fixes: 142288059
Test: existing
Change-Id: Iacf2b6685a153bdc51933e0b4eada00d36ae13ad