Prevent self causation issues on tests
... by generating ExternalSystemExceptions with a cause.

We currently simulate sync exceptions in tests by calling
registerSyncErrorToSimulate using an error message, this generates an
instance of ExternalSystemException with an empty cause (pointing to
itself). When getUserFriendlyError was called without a cause, this
would generate a Selfcasuation issue on the test exceptions.

This CL solves this by adding a Throwable as cause.

Test: Tests now work as expected.
Bug: 137231583
Change-Id: Ifa1c4b7e86e3e23775c8d57ad02eb96dea928e23