[Kotlin DSL Parser] Fix setting value for literals.
At the parser level, and when parsing kotlin() plugins or dependencies
we should be careful about how to set values of the created dsl
literals. The previousely used setValue() is for updating elements values that
will later be written back to the build file and involves creating a
psiElement to be used for updating the psiElement.
This CL moves the value setup logic to extractValue() that is invoked
at the literal creation time.

Bug: N/A
Test: existing tests already exist for value resolution.
Change-Id: Ifebb0b22120c3cb896994e5ebd4ac212adf0b4b8