Defaults AS to allow run app on multiple devices
The default behavior of IJ is to not allow a config to run
in parallel (i.e. more than one instance of the application
being run). This behavior isn't desired for Android Studio,
since it's very common for users to want to run the same
app on multiple devices. Furthermore, Android itself doesn't
allow the same config to run on the same device anyway.
Therefore, in all, it makes sense to default AS to allow
running the same app/config on multiple devices (except in
test cases, since that support still needs to be implemented).

Bug: 142086923
Test: manually verified new option is on by default and is saved
Change-Id: I1222497f23758f9763f497e0ce09afbfa0a989fd