[Dsl] rename model properties for AbstractFlavorType
The AbstractFlavorTypeDslElement is a base class for both BuildType
and ProductFlavor, capturing the behaviour (and naming the properties)
common to both subtypes.

Using the inheritance to do renaming, though, is a little painful: we
are already simulating multiple dispatch (on the Dsl element kind and
the external language), and it's not much fun to extend the simulation
to allow use of super in this effectively-multiple-dispatch world.

Instead, perform the effective inheritance statically: the property
model names are static, so when building the map we can concatenate
the abstract parent class's property information with the concrete
class's specific properties.  (This also means that all allocation is
done statically, which is probably a sensible design outcome.)

We do use the inheritance from the AbstractFlavor Dsl element to
handle special behaviour of certain operators (e.g. the appending
nature of proguardFiles).  This means that for now we have to do
element renaming after the call to super() in addParsedElement, not
before.  (In the longer term the behaviour of operators will be
table-driven, so there will be less -- ideally zero -- need to
special-case operators in add/setParsedElement() methods.)

As it happens, all the AbstractFlavor operators currently supported
have the same names in Groovy and Kotlin; nevertheless, define two
maps to allow for differences in currently-unsupported operators.

Test: existing
Bug: 141842964
Change-Id: Iee35b2833bc7b85fb4f3bf3a38b0bb5337bfba2a