Move get*SourceProviders methods to SourceProviderManager
and split their implementations between SourceProviderManagerImpl and
LegacySourceProviderManagerImpl.

This is a step in a series of similar steps aiming at making
IdeaSourceProvider a pure interface and moving it to project-system
module.

This change also simplifies some of the methods by removing unnecessary
branching depending on AndroidModel.isRequired() value which can be
expressed in a more general way.

This CL also moves Delegate and LegacyDelegate internal classes to
SourceProviderManager.kt.

NOTE: We implement new methods in SourceProviderManager by throwing
      UnsupportedException in order to avoid more changes this in this
      CL. This interface will be cleaned up later.

Bug: n/a
Test: existing
Change-Id: I1cc4a0d170769b981e771d2136310f2652d93cad