[Dsl] begin explicitly encoding Dsl semantics
Before this change, we have a limited, mostly implicit encoding of Dsl
semantics: one-argument method calls with names related to model
properties are assumed to be setter methods (with exceptions encoded
in addParsedElement override methods in DslElement classes);
assignments to names related to model properties likewise.  This fails
to capture various aspects of the Dsl semantics, such as implicit
conversions between types, and makes it difficult to handle semantic
differences or different API surfaces between Groovy and Kotlin Dsl
implementations.

This change begins making the encoding more explicit.  Generalize our
handling of names to encompass some idea of semantics, adapting the
maps in the various DslElement classes which implement renaming to
hold additional information: the maps, which prior to this encoded a
mapping of external to internal identifiers, become maps of (external
name, arity) pairs to (internal identifier, effect).

arity is:
- a non-negative number (a method taking exactly that many arguments);
- a negative number (a method taking at least as many arguments as the
  twos-complement of that number);
- null (a property, possibly read-only).

effect is:
- VAL: a read-only property;
- VAR: a read-write property;
- VWO: a write-only property;
- GET: a (0-argument) getter method;
- SET: a (1-argument) setter method;
- OTHER: any other method.

As of this CL, the arity and effect portions of the data are unused;
in fact the only visible effect of this CL should be to make the
lookup of external to model names very marginally slower, as we have
to iterate over the entries.  (This will be addressed in a follow-up
CL).

The order of defining the data is important, particularly for Kotlin
where we have VAL properties; we need to ensure when finding an
external name for a given property, that we produce something that
works with our existing constraints for creating new
elements (e.g. setDimension("xxx") rather than dimension("xxx") or
dimension = "xxx").  This will also be addressed in a follow-up CL.

Future CLs (or modifications of this CL) will:

- adapt creation to ignore the isMethod/asMethod part of the model,
  and instead search the map for (internal identifier, "VAR")
  and (internal identifier, "SET") values.  It might do this in
  different orders for Groovy and Kotlin: Groovy would prefer the
  "SET", and Kotlin the "VAR"...

- adapt resolution/extractValue to consider just values in the map
  with "VAR" or "VAL" semantics.  This will allow us to get
  b/142111082 right;

- handle adding properties with non-VAR/SET semantics uniformly:
  rather than implementing the same code three times for
  proguardFiles, consumerProguardFiles, resConfigs, write one
  APPENDN-handling routine.

- support getter methods in resolution/extractValue

- be able to suggest quickfixes based on the information in the
  model (e.g. if a user tries dimension = "xxx" in Kotlin, we can see
  that there is no VAR for that name but there is a VAL, and that the
  property with that VAL has a SET...)

- be able to detect properties being set in the wrong context, and
  offer quickfixes to move those properties to the right place

- (eventually) be able to remove the special-case parsing of
  one-argument method calls (assumed to be setter methods) which leads
  to problems elsewhere.

- probably other things too!

Test: existing tests pass
Bug: 142111082
Bug: 141842964
Bug: 72853928
Change-Id: Ibdd7a6ce3fea460c33a28c8767459dc18f633c44