Fix spurious error on preview initialization
When the project initializes and we do not have a gradle status yet, the
preview would display the render banner telling the user the project
needs to successfully build before the preview is usable. This is not
actually true and it's just the notification not taking into account the
null status from gradle.

Fixes: 142469859
Test: Manually checked
Change-Id: I56529ca9653f0cca5cded065c284439e54482de7