Make the editor to start in isRefreshing = true
When the editor is opened, by default was starting with isRefreshing =
false. This caused that if the editor started in dumb mode or if the
build was already done but we were waiting for the model, there in
notification but there is no preview either.

By starting the preview in the refreshing mode, the notification is
displayed fairly quickly so the user knows something is happening.

Fixes: 142674877
Test: Manually checked
Change-Id: I56735349723fd177d094b5ee757beb70454c0b18