[Res Explorer] Add Right click action to refresh a preview
Currently, there's no way to directly update the previews in the
Resource Explorer. So making changes to a Drawable or a Layout file will
not be reflected in the Resource Explorer, the preview can only be
updated indirectly by forcing it to display bigger preview, even in
those cases if the resource caused a preview error previously, that
error icon would never go away.

This adds an action to the right click menu called "Refresh Preview", it
clears the cached image for the selected resource which will cause the
preview to be generated again.

Supports multiple selection, as long as all selected resources can be
refreshed.

Bug: 137867724
Test: RefreshDesignAssetActionTest,
ResourceExplorerListViewModelImplTest
Change-Id: I91e7e67d46d7e3ec390549a187a0496b4e086456