Add in logic for the UserCounterDataSeries to display the event count.
Context: This data series will be used to show how many custom user events are occurring at a given time through a state chart in the monitor stage. Events are counted in 500ms intervals and added to the SeriesData accordingly.

Note: Fixed testing issue in CustomEventMonitor.kt that caused previous change to be reverted.

Bug: 140521019
Test: UserCounterDataSeriesTest.kt & CustomEventMonitor.kt
Change-Id: I903446c199fcd4cc6d484aef2df1498e2a5dc40e