Created module to move tests using previous AGP
Currently we have tests that depend on using an older version of AGP in
core tests, this causes all shards to have their own copy of each of
those versions since the whole intellij.android.core.tests module has a
dependency to those version, even if only some of the tests use them.

This CL adds a new module (intellij.android.old-agp-tests) that will
contain tests from core that depend on previous versions. Only a couple
of classes were moved to it at this moment, as a prrof of concept.

Bug: 141628806
Test: Moving tests, they should continue working as before this CL.
Change-Id: I49cc414b3a571e122f6d1b9a8701f3286fc9af47