Make executeCommand non-blocking
We added AppInspectionCommand.Id and AppInspectionEvent.Id in a separate
change.

We now generate our own version of commandId, and use it to map and
assign handlers. Because we no longer need to rely on commandId, we can
call transport's execute() asynchronously, and return from
executeCommand() right away. The "sending" is done in a dedicated
executor service.

Background:
Before, we had to wait for executeCommand to return. In testing
environment response Events are occasionally received before
execute() returns a commandId, leading to test failure.

Test: existing tests.
Fixes: 142762693, 142858738
Change-Id: I984e4ba78e0de8b3861db0cb68daf88e79536497