Stop Unit Tests connecting to external servers
When there is a dependency on a
early release for the Koltin pluging (eg 1.3.60-eap-23)
the wizard templates adds:
maven { url 'https://dl.bintray.com/kotlin/kotlin-eap' }
to build.gradle.
This url should be removed during unit testing,
otherwise the tests will connect to it when there
is a missing dependency.

Bug: n/a found locally, as a warning in bazel
Test: TemplateTest
Change-Id: If8f3617099e7d4ef69fd21798e11a345abce5ed4