[Dsl,.kts] use new GradleDslNamedDomain interfaces
The AbstractFlavorType (BuildType and ProductFlavor) and SigningConfig
Dsl elements are GradleDslNamedDomainElements.  Their respective
containers: BuildTypes, ProductFlavors and SigningConfigs, are
GradleDslNamedDomainContainers.

This allows rewriting the logic in the Kotlin writer to determine
which method name to use in terms of those interfaces: whether we
already have a method name, and if not, whether the element we are
about to create a reference to implicitly exists or not; the parser
also changes to use the same interface.

It also allows us to generalize the element deletion logic that checks
whether a method call should be deleted because it represents the
configuration of a Dsl element.

Test: existing (no change in functionality)
Bug: 141842964
Change-Id: I852d7993bafd9ad17ab8f8764e8018c488a67df9