[Dsl,.kts] parse configure-less NamedDomainObjects as block elements
A bare create("...") (without a { } configuration closure) should
still result in the Dsl containing a block element; configuration of
the element may happen later in arbitrary Dsl text.  Recognize that a
call expression with a 0-size lambdaArguments list is nevertheless a
candidate for being a block, and if so set the Psi element of the
resulting Dsl element to be the expression as a whole.

This allows stronger tests of element deletion.

Bug: 130224064
Test: strengthen testRemoveAndApplySigningConfig
Change-Id: I254e570523b17eed44121dbc9a022ba35f004ac0