[Asset Studio] Update the Android icon to match new branding
Manually updating the Android Vector Drawable to show the
updated branding (redesigned Adnroid head). The default icon in the
Vector/Image Asset clipart button and the 'android' icon in the
IconPickerDialog should now be the updated android branding.

This is while we await for the Google Fonts team to update the branding
on their side.

Fixes: 143077070
Test: n/a
Change-Id: I745f67e6833927b2b892b115970a43e378123207