Address exception in HostPanel.startLoading
HostPanel.startLoading launches a pooled thread, and the code in
that thread throws if the underlying XmlFile for the model is
null.

The model will throw this exception if the project has been disposed, which
seems like the most likely cause. This CL replaces the call to
NlModel.getFile (which verifies the file is not null) and calls
getPsiFileSafely instead, checking the return value. getPsiFileSafely
returns null if the project has been disposed, which can happen
since this code executes on a pooled thread.

Change-Id: Idf05e9408fabbb16ade9e68860aa9272d210b92e
Fixes: 140962832
Test: None