Remove dependency on AGP 2.2.0 for core tests
Tests that use it were moved to old-agp-tests. "2.2.0" can still be
found but only on stubs and unit tests that do not depend on having the
actual version available.

Now that version 2.2.0 some tests can be reenabled as well, this CL also
adds a test project for testLocalAarsAsModulesWithGradle2Dot2.

Bug: 141628806
Test: Manually run tests that have a 2.2.0 version in them from core.
Existing tests should cover this.

Change-Id: I17fe76bd14e88a705d913d529a00947dd4320b61