[Dsl] a sourceSet is a NamedDomainElement
... and the sourceSets element is a NamedDomainContainer.

It's a bit of a strange one, in that create() is not a
fully-functional interface: the method exists, but any attempt to use
it throws an error, either creating from an existing sourceSet or from
a failure to configure it properly, which implies that we should treat
all sourceSets as already existing.

The use of SingleArgumentMethodTransform in the root() SourceSet
property model implementation appears to have been an attempt to
handle multiple ways of expressing the same thing.  Unfortunately, the
implementation as it stands led to invalid output, e.g. in Groovy:

  setRoot setRoot('source')

which strengthening a test catches.  Removing the transform (and the
SingleArgumentMethodTransform constructor for which this was the last
use) improves the output, though this is an ideal case for handling
through table-driven Dsl semantics.

This has a side-effect of changing the order that source sets are
written when changed.  Adjust an existing test to be more forgiving of
the order of the sourceset elements in the parsed file.

Bug: 141842964
Bug: 142114586
Test: existing test strengthened, others updated
Change-Id: I2befd376103a194d0fa806e81106ea3164b034d5