Discard SourceProviderManager instance on configuration changes.
This is preparation allowing to simplify SourceProviderManager
implementations.  It allows implementations to cache configuration data
from the facet without the need to revalidate them each time the source
provider is used.

Future changes will rename to SourceProviderManager to SourceProviders
and further simplify implementations. The ultimate goal is to have
SourceProvider's classes be immutable data-like classes.

Bug: n/a
Test: unit test added (SourceProviderManagerBaseTest)
Change-Id: Iff888937dd3684ec06dced1e028ff9fcfa620a84