Add notification for Kotlin files with added previews
If a regular Kotlin file is added the preview import, we should open the
compose preview editor so the user can see the rendered previews. This
adds a notification that will allow the user to automatically open the
editor.

Also, add a single point where the default preview element finder is
configured so if we ever change it, we do not need to update every
constructor.

Fixes: 139463519
Test: Manually tested
Change-Id: I7fe548580a166c65ec7df2a28892ee73417fe23a