Remove incorrect usage of "Path"
Paths of files on Android device are not compatible with the JDK
"Path" class. Remove incorrect usage of the "Path" class and use
a "String" for these paths, which is consistent with other usages
in Device Explorer code.

A side effect of this change is that tests now pass on Windows.

Test: Existing
Bug: 141868211
Change-Id: Ie3298223cedba004135e00e4e699603cbf76fc3f