Don't call LaunchStatus.terminateLaunch from AndroidTestListener
This CL is a preliminary fix for supporting instrumentation tests
on multi-devices at a time.

Before this CL, AndroidTestListener calls terminateLaunch when
"am instrument" command is finished as a workaround solution to
forcefully terminate AndroidProcessHandler. This workaround is
needed because instrumentation test process is a short-lived
process (sometimes less than 50 ms) and AndroidProcessHandler
overlooks it (AndroidProcessHandler is polling live process
every second). Without the workaround, we ended up with
waiting for AndroidProcessHandler to find the already-finished
process until it times out.

Unfortunately, this CL doesn't solve the root problem but replaces
terminateLaunch() call with detachDevice() (See AndroidTestApplicationLaunchTask
and AndroidTestListener) so that AndroidProcessHandler may finish
before the timeout.

Unlike terminateLaunch(), this new workaround works fine with
multiple devices.

Bug: 143180401
Test: manual testing -- run instrumentation tests on multiple devices
Change-Id: I746afd1db8e0afd48ce9db629e2daf5d4f43a30e