Stop loading screen even when there are no new elements
In the case where there are no new elements but the loading screen is
showing we should also hide it once we are done with the refresh.
This prevents some cases where the loading screen would stay on forever
even when it is not really refreshing.

This would happen if the loading screen is shown after there was one
successful render.

Added additional logging and moved the notification update to the
updateSurfaceVisibilityAndNotifications method since we want to update
the notification bar in all cases we update the visibility.

Test: Manually tested
Fixes: 143174787
Change-Id: I80a3f56fde204d65d7d0d0af97f8439d3ee62d47