Update notifications to display only on the preview
This refactors the Compose Preview notification system to allow
notifications to only be displayed on the PreviewEditor side. This
allows for these notifications to be displayed in context and not
expanding the whole width of the FileEditor.

This CL adds a regular EditorNotification$Provider with the only tasks
of letting the PreviewEditor when the notifications should be updated.
This allows to replicating exactly the regular notification system and
the notification update calls.

When the PreviewEditor receives a call to update notifications, it will
call its own extension point and find providers that want to display in
the PreviewEditor.

Bug: 140732721
Test: Manually tested
Change-Id: I10c7e04de8fc72f06fb940735a9e312efdc2aa97