Handle connected virtual devices unknown by Studio
VirtualDevicesTask returns the virtual devices that Studio knows about and
displays in the Android Virtual Device Manager. I assumed that connected
virtual devices would always correspond to one of those virtual devices.
That assumption was bad. A developer can build their own system image and
run it in the emulator. The drop down did not display that connected virtual
device.

Bug: 137078988
Test: AsyncDevicesGetterTest::getImplDeveloperBuiltTheirOwnSystemImage
Change-Id: Idf766fc17ca6eb76f2b8ce09dafb2cd3c8653173