[Dsl] fix logic for when to stop deleting empty things
We want to terminate deletion of empty objects when the element we
would otherwise be about to delete is an empty container: an empty
map, or an empty list.  We detect this condition by checking the
container's Psi element for equality against the Psi element that we
are about to delete, except that because of operator precedence we
only performed this check in Kotlin for maps, and not for lists.

Although the situation in Groovy is different, it seems fairly likely
that the intent was that the equality check be performed for both
containers, so adjust this in Groovy as well: apparently no tests
excercised this, perhaps because of the structure of Groovy Psi (in
which the Psi objects representing elements of lists are direct
children of the Psi element representing the container; in Kotlin,
the tree has more depth).

Bug: 130224064
Test: new removeFromAndApplyListElements for Kotlin
Change-Id: I98c12aa90e5f2637f9939666ae53e786e348cf5a