Convert reselection to property change events.
The properties panel in the motion editor currently sends a new
selection event everytime a property value is changed. This causes
the property panel to recreate the UI elements.
In this CL detect selection changes for the same element and treat
these as property change events which simply updates the UI instead
of recreating it. As a side effect the focus will remain in the
currently edited property.

Bug: 142431987
Test: N/A
Change-Id: Iefb0bb4da222353b10a20af0d80c484c74e9f02e
(cherry picked from commit 698cddef406d776bd12b8bed9665ed7f0d89b483)