Fix how changes in binary resource files are handled
AndroidFileChangeListener had a special case for changes to images,
which didn't handle fonts. This was inconsistent with other code: adding
and removing binary files was handled in ResourceFolderRepository and
also supported fonts.

Modified the repository to correctly handle changes to all binary files and
removed the special case from AndroidFileChangeListener.

Test: ResourceFolderRepositoryTest
Bug: None
Change-Id: Ic85702a1362cdf8f83babaf9596dfeb7643b0da8