Register title actions with the model for updating
The actions in a title line in the properties panel should update
when the model is refreshed. Register the actions with the line
model such that this can be accomplished.
This will enable the actions in the Motion Editor to update after
property value changes have been detected.

Bug: 142431987
Test: N/A
Change-Id: I215469ede39feb9a21fea9a0b22ea3dca760a971
(cherry picked from commit 270bc90aee9660b6234e23dd9981286067056ff4)