Fix view z-order for real
The algorithm is now this:
Add the root view to level 0
recursively, add each child to the first level higher than the level of
its parent that contains no views intersecting with the current view.
(Prior to this change we just added to the first level with no
intersecting views, not the first level higher than the parent).

Test: added test in DeviceViewPanelModelTest
Bug: none
Change-Id: I7ce86040c644ef7bfbb129f70e86aaf631d51b90