Fix navigation in a combo box
Keyboard navigation in a combo box is disabled if the combo box
popup is not visible such that the keys can be forwarded to parent
components of the combo box for other purposes (such as scrolling
a scroll pane).
The code should check if the editor completions are visible as well.

Fixes: 143355242
Test: N/A
Change-Id: Idd2b9ad415d17bc89966b0e4439064c9e7340447
(cherry picked from commit 9b4b28367e2c59edfdf39aa21509e4c965177cd6)