[Dsl] tighten external names for property assignments
When deciding whether we are dealing with one of the known properties
for a given Dsl context: if we have an assignment, only canonize the
property with its internal (model) name if we have an entry in our
maps for a property (a null arity) whose semantics include write (VAR,
VWO).

Continue treating all method calls as candidates for renames
regardless of arity, except that we no longer consider property
entries as relevant.  This reveals a fair number of syntax errors in
the test data files.  Correct these tests.

Bug: 141548181
Bug: 141842964
Fixes: 142111082
Test: existing tests fixed
Change-Id: I679a5907c7592d8faefc9900439b939c0f3bc0d7