[Res Manager] Add copy and paste support for Kt files
Add basic support to 'Drag and Drop' and 'copy/paste' resources in to
code files.

For Kotlin if the paste operation is done on a method argument, it will
replace that argument with the selected resource.
If it's done on a method call, it will add the resource reference as one
of the arguments.

Otherwise it will just insert the string in to the position of the caret
or the drag and drop.

Note that this process is not very smart so it will not avoid producing
erronous results.

Bug: 142139061
Test: ResourcePasteProviderTest
Change-Id: Ie05e64fc8fd89018cd5f566783960244b81e826b