[Dsl] language-dependent syntax for added properties
We should never be renaming (in the sense of maybeUpdateName)
properties that exist in our model: that doesn't make sense.  Updating
names in this sense is for user-provided names (variables,
NamedDomainElements).  Replace the lookup of an external name from a
model name in maybeUpdateName with an assertion that the name is not
present.

The one remaining use of externalNameForParent() is the legitimate
one: where we're creating a new statement for a newly-created model
property.  In this case, it's not enough to return an external name,
though; we need to return an external name _and_ information about
whether it's a name suitable for assignment or not: the map of
external names to model properties is surjective but not
injective.  (NB: it's not even injective if you include the semantics:
there are multiple setter methods for most properties, though that's
less interesting).

Once we have the information about whether it's an assignment name or
a method name, we have the information we need at element creation
time to choose an appropriate name and mode, and we can for these
cases ignore the asMethod model property.  As proof-of-concept, have
Groovy prefer methods and Kotlin prefer assignments, adjusting
expected test output files to reflect this.

In the process, fix the renaming logic in the special handling of
testInstrumentationRunnerArguments: it was renaming to the model
property without canonizing.  ("Fix" by removing the call entirely and
letting setParsedElement() do the work).  Also adjust expectations in
tests in the difference between an empty manifestPlaceholders map and
no call at all.

Bug: 72853928
Bug: 141842964
Bug: 141970574
Bug: 142111082
Test: existing tests updated (often back to their original form)
Change-Id: Ia3870cc7c69df93c285e7bcc413725b6d5a0d085