[Res Explorer] Fix scroll jumping to next list of resources
When scrolling through resources the resources list would instantly
'jump' to the next section by scrolling so that the following section
would end up being at the top of the list. This behavior can be
frustrating since the user is not expecting the entire list to move when
they are scrolling through resources.

Bug: n/a
Test: manually
Change-Id: I488aa1ce9ad8dfccedf96dbb35c0dec06eb709e5