Turn AndroidFacetConfiguration.init into a static method in AndroidUtils
This way the logic for computing paths used by JPS can stay in
intellij.android.core but AndroidFacetConfiguration can be moved up to
intellij.android.common.

Facet editor tab creation was moved to an IntelliJ service. This UI is
only used in JPS projects and compiles against JPS APIs, so this way it
can stay in core, without the whole AndroidFacetConfiguration having to
be aware of JPS.

Bug: 143344535
Test: existing
Change-Id: Ice0f3da408a377a516c07f9e2e3948081909e5ad