Create a service for predefined custom cursors
The custom cursor cannot be created in headless environment, which
means the custom cursor cannot be tested by bazel. To solve that,
create a cursor provider interface and use it to get the custom
cursor. When testing, replacing the cursor provider implementation
and use the built-in cursors as expected output.

Bug: b/112468693, b/142953949
Test: updated and passed InteractionManager test
Change-Id: I53ff4dd5c14dd60b9b9336283a406b1dc12cbec4

GitOrigin-RevId: 752503d04b52840b997c16d0e652cc17c082b0ea