Add AppInspectionServiceRule to make testing easier
Primary motiviation is to make testing easier. The new rule takes care
of the setting up of the different pieces of transport and app
inspection service layer.

In addition, executor is hidden in AppInspectionTransport and is used for all
event callbacks now (used to be callbacks of
AppInspectionPipelineConnection were on direct executor.)

Fixed launchInspectorReturnsCorrectConnection test in
AppInspectionPipelineConnectionTest to be implementation agnostic.

Test: this is only refactoring
Fixes: 143642500, 143912268
Change-Id: I64d72e69226fb2e662596ea47a0153bd9a152320