[Res Picker] Support previews for Sample Data
Sample data doesn't follow the same structure of Drawable resources so
previewing them was not 'naturally' possible. They contain references to
multiple bitmap or xml files instead of just representing one of them.

This change makes it so that we can generate previews for SampleData by
making it preview the first image that the SampleData resource contains.
Eg: For @sample/background/scenic it previews
@sample/background/scenic[0].

Gradle-defined color resources are still not properly previewed.

Bug:133322931
Test: ResourceExplorerListVireModelImplTest
Change-Id: I14f33c3a21d3138761d23d38c5a5b117b83bb23c