Always issue START_ALLOC_TRACKING command
If a session is closed, and then a new session is
created on the same running process, memory profiler
isn't tracking allocations because memory profiler
doesn't issue START_ALLOC_TRACKING command.

Wny? When a new agent is connected, current implementation
doesn't issue START_ALLOC_TRACKING command if there's
an MEMORY_ALLOC_SAMPLING event in the session's range.

However, due to memory profiler's recent migration to
the transport pipeline, this is a bug because when the
old session is closed, a STOP_ALLOC_TRACKING command is
issued. So we must run START_ALLOC_TRACKING.

Bug: 143654550
Test: upadted MemoryProfilerTest#testLiveAllocationTrackingOnAgentAttach

Change-Id: If701592ef7fa2f97bb4af435b94ad23a3c06a3b5