Move facet.isDisposed checks from SourceProviders
Instead, move them to the facet usage site.

The isDisposed checks in SourceProviders was added as a workaround and
it does not solve the problem but sometimes hides it. This CL moves the
checks to the usage site where it is not cleat whether the facet can be
already disposed. The places where the facet used is freshly obtained
are left intact.

This is required to decouple the implementation of SourceProviders from
facets and models.

Bug: n/a
Test: existing
Change-Id: Ifc80b4169c24b75508576f0fb323077e868f5e71