143867725: Perform import at the same time as Add Dependency
When the Kotlin or Java editors show a class symbol as unresolved, we
add an intention action to import the associated Maven library and
perform a sync.

However, after invoking the import action, the symbol stays red since
the sync usually takes a while. And even when the sync is done, the
symbol stays red because you haven't actually imported the class.

This change improves this situation. It stores the associated fully
qualified name in our symbol to library lookup table, and when you
invoke the action to import the library, we also import the
symbol. That way, as soon as the sync finishes, the symbol is resolved
as well.

This CL also fixes the following bug:
143867090: Unresolved library import isn't offered on package names

Test: AndroidMavenImportIntentionActionTest, which has been updated
Fixes: 143867725
Fixes: 143867090
Change-Id: I625a55236c1af3999a71737d99b30b10abc7ccf4