Replace custom ActionManager with a custom toolbar
ActionManager allows to customize the toolbar for different file
selections while Toolbar action groups provide toolbar per file
type, which we need here. So this is the change to use a more
tailored tool for the problem. This also makes custom veiw and
compose preview more similar which simplifies the merge.

Bug: 143067434
Test: all tests shoudl pass
Change-Id: I11327818ad69348dcee3b13a8a147f5a574dbba8