Hide "Hide preview" checkbox if split editor is enabled
Users can opt out of hiding the preview window when editing non-layout
files. This only makes sense if we're not using the split editor, since
we never show the window in this case. Therefore, we should hide this
option when the split editor flag is enabled.

This CL also merges the UI designer form, now discouraged, into the
panel code and takes the opportunity to convert it to kotlin to make
advantage of IntelliJ UI DSL.

Fixes: 143869159
Test: Manually verified by going to Settings -> Editor -> Layout Editor
Change-Id: I4d2a5c2e9476a3ad92f8bb1986fa4574e7956aee