[KTS] Move getGradle*File util methods to sdktools.common
This change, along with the other one in this topic moves the
getGradleBuildFile and getGradleSettingsFile util methods to
sdktools.common. These util methods find a .gradle or .gradle.kts build
or settings file in a given directory. This funtionality exists in a
bunch of place e.g lint, IDE etc. The move allows the sharing of this
method.

Test: Util tests also moved
Bug: N/A
Change-Id: I359b459b84cff1802b4b6dc429adf09ab7f85e8d