Add POM to SourcesAndJavadocArtifact model.
IDE finds the location of POM file based on the path of
library jar/aar, assuming the location of POM is related to
jar/aar. The assumptions includes they're placed in the same
folder, or under the same parent's parent folder.
This sometimes fails to cover all cases based on the version
of Gradle.
A planned change on AGP to make library path return processed
jars will make this guessing process impossible.

This CL quries the exact location of POM file in Gradle
so that IDE will not need to rely on locaiton of jar/aar.

Bug: 136157081
Test: integration test added in next cl.
Change-Id: I1e57093d21d7a871d40b65124fff016e6baad061