Short-circuit accept method condition to prevent memory leak
We currently have a workaround solution to resolve conflicts in files
that contain both @Composable and Android Custom Views. The provider of
custom views, CustomViewEditorProvider, only accepts files that are not
accepted by ComposeFileEditorProvider. The problem is the former creates
an instance of the latter every time it's called, causing a memory leak
since the providers indirectly reference DesignerTypeRegistrar#types.

This CL changes the order of CustomViewEditorProvider#accept condition
to short-circuit if the NELE_CUSTOM_VIEW_PREVIEW flag is disabled,
preventing the leak from happening in this case.

Bug: 143067434
Test: NavEditorMemoryUseTest
Change-Id: I99941a0705736520c731788658623ce4ac43b7f0