Code cleanup (fixmes): revert android-kotlin module to its state "before merge" (aosp/AndroidStudio 3.5)
Currently IJ uses precompiled version of the android-kotlin module. Remove all the "fixmes" introduced by the merge AndroidStudio3.5->IJ and revert the code to its state in AOSP repository (in order to not introduce merge conflicts in future).

There is an intention (in future) to include android-kotlin module in IJ as sources.

GitOrigin-RevId: 45e7151c74af6c5d47c33ea86185e43f11bf0175