Fix crash in NavigationSchema.buildTypeToDefaultTag
If the user replaces navigation-fragment with
navigation-runtime in the gradle dependencies,
we can pass our validation in NavigationSchema.init()
but then throw an exception in buildTypeToDefaultTag
because the fragment navigator is not there. Add null checks
to make the code more tolerant.

Change-Id: Ic76e547b429bb3f24a351956c3477615761d3a9c
Fixes: 143898223
Test: None
(cherry picked from commit 1194491766766f836915561d7cd9ca20d3c64964)