[Dsl] generalize AaptOptions support
Implement name mapping and semantics tables for the aaptOptions Dsl
block.  This brings to the fore syntax errors in our Kotlinscript test
files; fix those appropriately.

The ignoreAssets and ignoreAssetsPattern Dsl operators and properties
all affect a single model-internal property, arbitrarily chosen here
as IGNORE_ASSETS.  Simplify the implementation of the ignoreAssets()
model, and prefer the -Pattern suffixed operators for externalization
because of the historical setIgnoreAssetsPattern operator.

Bug: 130224064
Bug: 141842964
Test: existing tests updated
Change-Id: Icc0005e0d1336a34d3480fc7b7a2b838c2e7e268