Use SourcesAndJavadocArtifact model to find pom file.
IDE finds the location of POM file based on the path of
library jar/aar, assuming the location of POM is related to
jar/aar. The assumptions includes they're placed in the same
folder, or under the same grandparent's folder.
This sometimes fails to cover all cases based on the version
of Gradle.
A planned change on AGP to make library path return processed
jars will make this guessing process impossible.

This CL uses SourcesAndJavadocArtifact model to get the location
of POM, which finds the exact location of POM file from Gradle.

The old mechanism of locating pom file is still used as fall back
and will be removed when SourcesAndJavadocArtifact is available
to Java modules.

Bug: 136157081
Test: included.
Change-Id: I4cb69e6e01fef62a6d30e3fb77ad9d7ad508f9ee