[Dsl] start removing asMethod / isMethod
Dsl elements which have external-to-model maps do not use the model
isMethod property to determine whether to write out method calls or
assignments, but instead use the maps in an external-language-specific
way.  This means we can remove the .asMethod() calls to
GradlePropertyModelBuilders (whether implicit or explicit) in building
models for those elements.

Bug: 141970574
Test: existing (no change in functionality)
Change-Id: I342ec1546a9300890a387bb711de4e6598b44041