Make BLeak options more easily customizable
As a step towards the extraction of BLeak into a standalone unit, this
change consolidates the specification of various parameters (e.g.
whitelists, which custom Expanders to use, iteration count, etc.) into
the BleakOptions class, an instance of which is passed to runWithBleak.

Whitelists and leaktraces are also given a more explicitly typed
representation, which should make them easier to use from Java.

Issue: 143490763
Change-Id: I205546eaa926e92a17f6efbebc9e33e2dba21720