Fix BuildAttributionTreeView instances being leaked
BuildAttributionManagerImpl creates an instance of BuildAttributionTreeView
after building, and the latter registers itself in the Disposable tree as a
children of the current Project. This means we can only dispose this object
when the project is disposed, leaking all the instances created meanwhile.

This CL fixes the issue by registering the BuildAttributionTreeView as a
children of the ContentImpl object that displays it, which in turn is
registered as a children of the Project and will be disposed in case we
dispose the Project. As the ContentImpl is added to the BuildContentManager,
it will also be disposed when the content manager is disposed, consequently
disposing the BuildAttributionTreeView that won't be leaked anymore.

Bug: 144178249
Test: Leak exposed by a BLeak test introduced in change I84ffdf8, which
      passes after this change. _LastInSuiteTest#testProjectLeak should
      also pass.

Change-Id: I95e770b37267c79ee09d30b2e177fc32a410f91f