[Dsl,.kts] fix convertReferenceText crash
In KotlinScript, convertReferenceText requires creating and inspecting
Psi in order to convert Kotlin syntax to the dotted-notation that the
Dsl requires.  Do that under runReadAction explicitly, because there
are callers of resolveReference (the only caller of
convertReferenceText) which do not already have a read lock.

Also, mark the PsiFactory results as not being inserted into a source
file: we are simply going to walk the Psi to generate a string.  This
should save time and memory.

(It is a little ugly to be creating Psi, particularly since it is
presumably Psi that we already have from an existing parsed file.  We
would still need the readAction even without that.)

Bug: 144230634
Test: should remove this crash from crashlogs
Change-Id: Ie4431e4b9944a58e3098fed3adc373847df3f010